+++ src/components/sidebar/Sidebar.spec.jsx
@@ -80,7 +80,7 @@ const server = setupServer(
   })
 )
 
-beforeAll(() => server.listen())
+beforeAll(() => server.listen({ onUnhandledRequest: 'bypass' }))
 afterEach(() => server.resetHandlers())
 afterAll(() => server.close())
 
@@ -100,6 +100,7 @@ describe('Sidebar Component', () => {
         id: 1,
       },
     },
+    lobby_id: 1,
     status: 'online',
   }